+++ database/migrations/2021_04_14_140151_create_countries_table.php
@@ -35,6 +35,8 @@ class CreateCountriesTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('countries');
+        if (Schema::hasTable('countries')) {
+            Schema::dropIfExists('countries');
+        }
     }
 }
