+++ database/migrations/2021_04_14_140216_create_states_table.php
@@ -32,6 +32,8 @@ class CreateStatesTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('states');
+        if (Schema::hasTable('states')) {
+            Schema::dropIfExists('states');
+        }
     }
 }
